restructure test suite in internal/docker
- Consolidate all parser tests into parser_test.go (moved TestParseComposePSJSON)
- Create docker_test.go for all Client method tests (consolidated from compose_project_test.go and network_test.go)
- Add new container_test.go with comprehensive tests for Container struct methods
- Remove redundant test files (compose_project_test.go, network_test.go)

This improves test organization by aligning test files with their corresponding source files:
- parser_test.go tests parser.go functions
- docker_test.go tests docker.go Client methods
- container_test.go tests container.go Container methods
