--- internal/docker/compose_project_test.go
@@ -1,17 +1,0 @@
-package docker
-
-import (
-	"testing"
-)
-
-func TestListProjects(t *testing.T) {
-	// This is a basic test that just checks the method exists
-	// In a real test environment, you would mock the exec.Command
-	client := NewClient()
-
-	// We can't actually test listing projects without docker compose
-	// The method should return an error or empty result
-	_, err := client.ListComposeProjects()
-	// Either error or empty result is acceptable
-	_ = err
-}
