--- internal/docker/network_test.go
@@ -1,64 +1,0 @@
-package docker
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestClient_ListNetworks(t *testing.T) {
-	// Test data from actual docker network ls --format json output
-	testCases := []struct {
-		name           string
-		mockOutput     string
-		expectedCount  int
-		expectedFirst  string
-		expectedDriver string
-	}{
-		{
-			name: "parse docker network ls JSON output",
-			mockOutput: `{"CreatedAt":"2025-06-25 07:13:05.579113899 +0900 JST","Driver":"bridge","ID":"58c772315063","IPv4":"true","IPv6":"false","Internal":"false","Labels":"com.docker.compose.version=2.33.1,com.docker.compose.network=default","Name":"blog4_default","Scope":"local"}
-{"CreatedAt":"2025-06-25 07:10:51.709317902 +0900 JST","Driver":"bridge","ID":"0e4dfb157d47","IPv4":"true","IPv6":"false","Internal":"false","Labels":"","Name":"bridge","Scope":"local"}
-{"CreatedAt":"2025-06-22 21:41:39.32360742 +0900 JST","Driver":"host","ID":"4dc1e78a8da8","IPv4":"true","IPv6":"false","Internal":"false","Labels":"","Name":"host","Scope":"local"}`,
-			expectedCount:  3,
-			expectedFirst:  "blog4_default",
-			expectedDriver: "bridge",
-		},
-		{
-			name:          "empty output",
-			mockOutput:    "",
-			expectedCount: 0,
-		},
-		{
-			name:           "single network",
-			mockOutput:     `{"CreatedAt":"2025-06-25 07:10:51.709317902 +0900 JST","Driver":"bridge","ID":"0e4dfb157d47","IPv4":"true","IPv6":"false","Internal":"false","Labels":"","Name":"bridge","Scope":"local"}`,
-			expectedCount:  1,
-			expectedFirst:  "bridge",
-			expectedDriver: "bridge",
-		},
-		{
-			name:           "network with internal=true",
-			mockOutput:     `{"CreatedAt":"2025-08-03 11:10:15.441860358 +0900 JST","Driver":"bridge","ID":"c65e5d9416a7","IPv4":"true","IPv6":"false","Internal":"true","Labels":"com.docker.compose.network=development","Name":"dcv-development","Scope":"local"}`,
-			expectedCount:  1,
-			expectedFirst:  "dcv-development",
-			expectedDriver: "bridge",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Note: This is a unit test that would need mocking
-			// For now, we're testing the parsing logic by directly testing
-			// the models conversion
-
-			// Since we can't easily mock ExecuteCaptured, let's at least
-			// verify the parsing logic works with the expected format
-			require.NotNil(t, tc.mockOutput, "Test case setup validation")
-
-			// Test that we can create a client
-			client := NewClient()
-			assert.NotNil(t, client)
-		})
-	}
-}
